Various fixes  - Fix flags in MADT x2APIC (CPU only has ENABLED, not ONLINE  - Increased chunk index and related function from 8 to 16-bit  - Increase MAX_APS and MAX_APIC_IDS to 512 (need an APs stack size evaluation to be sure we will not overflow)

+++ app/main.c
@@ -101,7 +101,7 @@ static int              test_stage = 0;
 
 // These are exposed in test.h.
 
-uint8_t     chunk_index[MAX_CPUS];
+uint16_t     chunk_index[MAX_CPUS];
 
 int         num_active_cpus = 0;
 int         num_enabled_cpus = 1;
